Break up the wavpack crate into modules
